resolve smack permission issuses and update ttrace to the latest version * remarkable changes - creates ttrace_tag file when boot-up the device - provides static lib. version of ttrace library - dynamically generates ttrace.h

+++ include/TTraceWrapper.h
@@ -22,11 +22,13 @@
 
 #include <ttrace.h>
 #include <stdarg.h>
+#include <stdint.h>
+
 class TTraceWrapper {
 	private:
-		int tag;
+		uint64_t tag;
 	public:
-		TTraceWrapper(int tags, const char* label, ...);
+		TTraceWrapper(uint64_t tags, const char* label, ...);
 		~TTraceWrapper();
 };
